Move fixture to contracts_v1.py
