+++ starknet_py/tests/e2e/fixtures/contracts.py
@@ -1,20 +1,13 @@
 # pylint: disable=redefined-outer-name
-import json
-from typing import Any, Dict, List, Optional, Tuple
+from typing import List
 
 import pytest
 import pytest_asyncio
 
-from starknet_py.common import (
-    create_casm_class,
-    create_compiled_contract,
-    create_sierra_compiled_contract,
-)
+from starknet_py.common import create_compiled_contract
 from starknet_py.constants import FEE_CONTRACT_ADDRESS
 from starknet_py.contract import Contract
-from starknet_py.hash.casm_class_hash import compute_casm_class_hash
 from starknet_py.net.account.base_account import BaseAccount
-from starknet_py.net.udc_deployer.deployer import Deployer
 from starknet_py.tests.e2e.fixtures.constants import (
     CONTRACTS_COMPILED_DIR,
     CONTRACTS_DIR,
@@ -37,19 +30,6 @@ def map_compiled_contract() -> str:
     Returns compiled map contract.
     """
     return read_contract("map_compiled.json", directory=CONTRACTS_COMPILED_DIR)
-
-
-@pytest.fixture(scope="package")
-def sierra_minimal_compiled_contract_and_class_hash() -> Tuple[str, int]:
-    """
-    Returns minimal contract compiled to sierra and its compiled class hash.
-    """
-    compiled_contract = read_contract("minimal_contract_compiled.json", v1_v2=True)
-    compiled_contract_casm = read_contract("minimal_contract_compiled.casm", v1_v2=True)
-    return (
-        compiled_contract,
-        compute_casm_class_hash(create_casm_class(compiled_contract_casm)),
-    )
 
 
 @pytest.fixture(scope="package")
@@ -99,40 +79,6 @@ async def deploy_contract(account: BaseAccount, class_hash: int, abi: List) -> C
     )
     deployment_result = await deployment_result.wait_for_acceptance()
     return deployment_result.deployed_contract
-
-
-async def deploy_v1_contract(
-    account: BaseAccount,
-    contract_file_name: str,
-    class_hash: int,
-    calldata: Optional[Dict[str, Any]] = None,
-) -> Contract:
-    """
-    Deploys Cairo1.0 contract.
-
-    :param account: An account which will be used to deploy the Contract.
-    :param contract_file_name: Name of the file with code (e.g. `erc20` if filename is `erc20.cairo`).
-    :param class_hash: class_hash of the contract to be deployed.
-    :param calldata: Dict with constructor arguments (can be empty).
-    :returns: Instance of the deployed contract.
-    """
-    contract_sierra = read_contract(contract_file_name + "_compiled.json")
-    sierra_compiled_contract = create_sierra_compiled_contract(
-        compiled_contract=contract_sierra
-    )
-    abi = json.loads(sierra_compiled_contract.abi)
-
-    deployer = Deployer()
-    deploy_call, address = deployer.create_contract_deployment(
-        class_hash=class_hash,
-        abi=abi,
-        calldata=calldata,
-        cairo_version=1,
-    )
-    res = await account.execute(calls=deploy_call, max_fee=MAX_FEE)
-    await account.client.wait_for_tx(res.transaction_hash)
-
-    return Contract(address, abi, provider=account, cairo_version=1)
 
 
 @pytest_asyncio.fixture(scope="package")
